backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow<3.20
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
bcrypt==4.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
alembic==1.12.1
Flask-Migrate==4.0.5